+++ shaders/gles20/posFS.glsl
@@ -1,5 +1,5 @@
-precision highp float;
-precision highp sampler2D;
+precision mediump float;
+precision mediump sampler2D;
 
 uniform sampler2D uPosTex;
 uniform sampler2D uVelTex;
